Changing variable type to decimal

+++ BudgedManager/Models/Entity/Limit.cs
@@ -8,8 +8,8 @@ public class Limit
     public int CategoryId { get; set; }
     
     [Range(0, int.MaxValue)]
-    public int LimitWarning { get; set; } //set LimitAlert = Limit * 0.9 for 90% of Limit
+    public decimal LimitWarning { get; set; } //set LimitAlert = Limit * 0.9 for 90% of Limit
     
     [Range(0, int.MaxValue)]
-    public int LimitAlert { get; set; }
+    public decimal LimitAlert { get; set; }
 }
